WIP: Move Injector and SDS classes to proper files
All other jet_tracking classes will remain in jet_tracking's
devices.py file as they are specific to that application.
